requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2023.3
sqlparse==0.4.3
